+++ src/Page/Admin/Store/StoreUpdatePage.tsx
@@ -1,6 +1,6 @@
 import { useForm } from "react-hook-form";
 
-import { InputDefault, LabelDefault } from "@/Components/UI/Atoms";
+import { ErrorLabel, Form } from "@/Components/UI/Atoms";
 import { Loading, StoreUpdateStatusBar } from "@/Components/UI/Molecules";
 import {
   useGetStore,
@@ -8,7 +8,6 @@ import {
   useUpdateStore
 } from "@/Page/Admin/hooks";
 import { IStoreFormProps } from "../interface";
-import { Form, InputContainer } from "./styles";
 
 const StoreUpdatePage = () => {
   const { data: updateStore, isLoading: isLoadingGetStore } = useGetStore();
@@ -48,19 +47,23 @@ const StoreUpdatePage = () => {
       ) : (
         <>
           <Form onSubmit={onSubmit}>
-            <InputContainer disabled>
-              <LabelDefault htmlFor="code">사업자번호</LabelDefault>
-              <InputDefault
+            <Form.FormItemContainer>
+              <Form.Label htmlFor="code">사업자번호</Form.Label>
+              <Form.Input
                 id="code"
                 placeholder="사업자 번호를 입력해주세요."
                 {...register("code", {
+                  required: "사업자 번호가 꼭 필요합니다.",
                   value: updateStore?.code
                 })}
               />
-            </InputContainer>
-            <InputContainer>
-              <LabelDefault htmlFor="name">가게이름</LabelDefault>
-              <InputDefault
+            </Form.FormItemContainer>
+
+            <ErrorLabel>{errors.code?.message}</ErrorLabel>
+
+            <Form.FormItemContainer>
+              <Form.Label htmlFor="name">가게이름</Form.Label>
+              <Form.Input
                 id="name"
                 placeholder="가게 이름을 입력해주세요."
                 {...register("name", {
@@ -68,10 +71,12 @@ const StoreUpdatePage = () => {
                   required: "가게 이름이 꼭 있어야합니다."
                 })}
               />
-            </InputContainer>
-            <InputContainer>
-              <LabelDefault htmlFor="address">주소</LabelDefault>
-              <InputDefault
+            </Form.FormItemContainer>
+            <ErrorLabel>{errors.name?.message}</ErrorLabel>
+
+            <Form.FormItemContainer>
+              <Form.Label htmlFor="address">주소</Form.Label>
+              <Form.Input
                 id="address"
                 placeholder="주소를 입력해주세요."
                 {...register("address", {
@@ -79,10 +84,12 @@ const StoreUpdatePage = () => {
                   required: "주소가 꼭 있어야합니다."
                 })}
               />
-            </InputContainer>
-            <InputContainer>
-              <LabelDefault htmlFor="phone">전화번호</LabelDefault>
-              <InputDefault
+            </Form.FormItemContainer>
+            <ErrorLabel>{errors.address?.message}</ErrorLabel>
+
+            <Form.FormItemContainer>
+              <Form.Label htmlFor="phone">전화번호</Form.Label>
+              <Form.Input
                 id="phone"
                 placeholder="가게 대표 번호를 입력해주세요."
                 {...register("phone", {
@@ -90,13 +97,12 @@ const StoreUpdatePage = () => {
                   required: "대표 번호가 꼭 필요합니다."
                 })}
               />
-            </InputContainer>
+            </Form.FormItemContainer>
+            <ErrorLabel>{errors.phone?.message}</ErrorLabel>
+
             <input style={{ visibility: "hidden" }} type="submit" />
           </Form>
-          <LabelDefault className="error-label">
-            {errors.addFail?.message}
-          </LabelDefault>
-
+          <ErrorLabel>{errors.addFail?.message}</ErrorLabel>
           <StoreUpdateStatusBar onSubmit={onSubmit} />
         </>
       )}
