+++ src/Page/Admin/Store/styles.tsx
@@ -67,11 +67,7 @@ export const InputContainer = styled.div<{ disabled?: boolean }>`
     border: unset;
     outline: unset;
   }
-  .error-label {
-    grid-column: 2 / 10;
-    ${SubTitle2};
-    color: ${(props) => props.theme.color.error500};
-  }
+
   ${({ theme }) => theme.device.tablet} {
     padding: 0.8rem 0;
     label {
